# test(e2e): split e2e test setup phase and separated IBC setup