--- a/tests/e2e/e2e_test.go
+++ b/tests/e2e/e2e_test.go
@@ -64,6 +64,7 @@ func (s *IntegrationTestSuite) TestGov() {
 }
 
 func (s *IntegrationTestSuite) TestIBC() {
+	s.ensureIBCSetup()
 	if !runIBCTest {
 		s.T().Skip()
 	}