## Description

This PR splits the setup and teardown of the e2e tests into phases. The first setup phase is executed at the beginning of any e2e test and initializes only one chain, without setting up the secondary chain and IBC. The second phase is lazily executed when IBC is needed and performs the IBC setup. 
For most of the tests IBC is not required, and this PR gives substantial speedup on single e2e tests (see below).

The only additional requirement for this PR to work is that any tests that requires IBC to be setup, should start with a call to `s.ensureIBCSetup()`.

Split Setup Phases times:
All e2e tests: 390s
TestPhoton e2e tests: 13s

Current e2e times:
All e2e tests: 469s
TestPhoton e2e tests: 94s